api/requirements.txt
attrs==21.4.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
ecdsa==0.17.0
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-HTTPAuth==4.5.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
pytest==7.0.0
requests==2.27.1
rsa==4.8
six==1.16.0
tomli==2.0.1
urllib3==1.26.8
uWSGI==2.0.20
Werkzeug==2.0.2
